implement accesstokens page and remove accesstokens from toppage
